requirements.txt
asgiref==3.4.1
Django==3.2.11
django-cors-headers==3.10.1
django-filter==21.1
django-webpack-loader==1.4.1
djangorestframework==3.13.1
Markdown==3.3.6
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
